backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
"bcrypt==3.2.0"